+++ HISTORY.md
@@ -7,6 +7,8 @@
 - Implemented `vhpi_get_phys`
 - Added command `--list` to print all units in a library
 - Rewrote the `run_regr` program in C so `make check` no longer requires Ruby
+- The `--relax=generic-static` option is now renamed to `--relax=locally-static` and
+  applies in wider range of contexts
 
 ## 1.0 - 2015-05-01
 - First stable release